fixes to syntax, size and colors of components

+++ src/pages/LoginPage.js
@@ -3,9 +3,21 @@ import LoginForm from "../components/LoginForm/LoginForm";
 import "../main.css";
 import Template from "../components/Template/Template";
 import { deleteTokenLocalStorage } from "../helpers/userHelper";
+import { useLogedUser } from "../context/UserContext";
 
 const LoginPage = () => {
-  deleteTokenLocalStorage();
+  const { user, loading } = useLogedUser();
+  const isAuth = user?.token;
+
+  if (loading) {
+    return <h1>Loading..</h1>;
+  } else {
+    if (isAuth) {
+      deleteTokenLocalStorage();
+      window.location.reload();
+    }
+  }
+
   return (
     <Template>
       <LoginForm />
